created taskmaster class to handle the scrolling of objects, might not need to be a class, but allows for easy changes off of main.
changed some things in the windowcapture class to allow the cropping and offset of the screen recording to be changed instead of hard coded. this will allow for the creation of an additional readimage object to look at the page number and determine when to stop.

i am not happy with the functionality of the pyautogui module. it feels very limited, and i will likely change it in the future
the scroll functionality is not very usable for this task, and the dragto functionality triggers the continious dragging of the page, also not ideal.

however the proof of concept has been made, now to make it more accurate when i next have a battle log availabe and then implement the ability to click through pages